Split listener to serveral classes
