+++ wsmaster/codenvy-hosted-api-workspace/src/test/java/com/codenvy/api/workspace/server/jpa/cascaderemoval/JpaEntitiesCascadeRemovalTest.java
@@ -19,7 +19,8 @@ import com.codenvy.api.permission.server.spi.PermissionsDao;
 import com.codenvy.api.workspace.server.jpa.JpaRecipePermissionsDao;
 import com.codenvy.api.workspace.server.jpa.JpaStackPermissionsDao;
 import com.codenvy.api.workspace.server.jpa.PermissionsJpaModule;
-import com.codenvy.api.workspace.server.jpa.RemovePermissionsOnLastUserRemovedEventSubscriber;
+import com.codenvy.api.workspace.server.jpa.listener.RemovePermissionsOnLastUserRemovedEventSubscriber;
+import com.codenvy.api.workspace.server.jpa.listener.RemoveStackOnLastUserRemovedEventSubscriber;
 import com.codenvy.api.workspace.server.recipe.RecipeDomain;
 import com.codenvy.api.workspace.server.recipe.RecipePermissionsImpl;
 import com.codenvy.api.workspace.server.spi.WorkerDao;
@@ -277,7 +278,7 @@ public class JpaEntitiesCascadeRemovalTest {
     @DataProvider(name = "beforeRemoveRollbackActions")
     public Object[][] beforeRemoveActions() {
         return new Class[][]{
-                {RemovePermissionsOnLastUserRemovedEventSubscriber.class}
+                {RemoveStackOnLastUserRemovedEventSubscriber.class}
         };
     }
